Development (#180)
* FIX: not needed notification on model load/perspective change
FIX: Expressions for OTRel of soft goals

* FIX: Configuration operation working with new attributes.

* FIX: Define all soft goals related attributes with the same domain
(0..4)

* FIX: Order corrected for new simulation
FIX: Broken models avoided when a clone is lost and the clonned element
has relations. The element must be deleted and the relations created
again, but the model is not broken.

* FIX: constructor for RIGHTCONCEPTVARIABLE with correct semantic concept
FIX: RIGHTVARIABLE replaced by RIGHTCONCEPTVARIABLE for refas opers meta
model
FIX: expressions for soft goals old implementation fixed. New
implementation still incomplete

* FIX: sg to sg otrel working for dynamic operations (not optimal, not
suitable for MUTEX or RANGE): works for AND and OR

* FIX: use of OTRel and PWRel types visually defined. DisplayName and
Panel name not working.

* NEWFUN: Include tooltip for attribute's labels

* FIX: some tooltips included

* FIX: Required Level working for sg in static simulation. Included in
domain values for ClaimExpLevel

* FIX: duplicated relation type removed