Refactor 1.2.0.1. Place Load file action  into the class VariamosGuiEditorActions and VariamosGuiEditorActionsController. Both classes were documented.
Remove old methods for opening new files

Add inside variamosGraphEditor a new variable to save the last path
where a filechooser is open. Doing so, each time a new file is open, it
is possible to open it from the last visited path, while before this fix
each time a filechooser was open, the user directory was used as path by
default.